Survey data from the Alaska and Northwest Fishery Science Centers

Various processed survey datasets.

biomass

parameters

specimens

length_comps

age_comps

Format

biomass: Design-based biomass estimates. A data frame.

science_center

Name of the U.S. Science Center.

region

U.S. region.

area

Survey name indicating the region of the data.

common_name

Species common name.

scientific_name

Species scientific name.

year

Year.

est

Mean biomass estimate.

lwr

Lower bound of the biomass estimated base on 95% confidence interval.

uwr

Upper bound of the biomass estimated base on 95% confidence interval.

parameters: Estimated growth parameters. A list of two data frames.

region

U.S. region.

common_name

Species common name.

sex

Sex (Female, Male, Unsexed).

k

Estimated von Bertanlaffy growth rate.

Linf

Estimated maximimum size in cm.

L0

Estimated minimum size at in cm.

CV1

Estimated coefficient of variation around L0.

CV2

Estimated coefficient of variation around Linf.

median_intercept

Median values for the length-mass intercept parameter (A).

SD

Estimated standard deviation for length-mass relationship.

A

Length-mass intercept A (A * length_cm ^ B).

B

Length-mass exponential parameter B (A * length_cm ^ B).

specimens: Biological data. A data frame.

science_center

Name of the U.S. Science Center.

region

U.S. region.

survey_id

Survey name indicating the region of the data.

common_name

Species common name.

scientific_name

Species scientific name.

year

Year.

sex

Sex (Female, Male, Unsexed).

length_cm

Length of fish in cm.

age_years

Age of fish in year.

mass_kg

Mass of fish in kg.

length_comps: Expanded length comps. A data frame.

science_center

Name of the U.S. Science Center.

region

U.S. region.

area

Survey name indicating the region of the data.

common_name

Species common name.

scientific_name

Species scientific name.

year

Year.

sex

Sex (Female, Male, Unsexed).

length_cm

Length of fish in cm.

population_count

Expanded observations by length.

total

Sum of the expanded length observations by year across all sexes.

propotion

Proportion of observations by length and sex that sums to 100 for each year across all observed sexes.

age_comps: Expanded age comps. A data frame.

science_center

Name of the U.S. Science Center.

region

U.S. region.

area

Survey name indicating the region of the data.

common_name

Species common name.

scientific_name

Species scientific name.

year

Year.

sex

Sex (Female, Male, Unsexed).

age_years

Age of the fish in years.

population_count

Expanded observations by age.

total

Sum of the expanded age observations by year across all sexes.

propotion

Proportion of observations by age and sex that sums to 100 for each year across all observed sexes.